# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "BPMT API's",
        "description": "Project management operations",
        "version": "1.0.0"
    },
    "host": "http://my-default-host.com",
    "basePath": "/api/",
    "schemes": [
        "http",
        "https"
    ],
    "paths": {
...
        "/v1/sprint": {
            "post": {
                "tags": [
                    "Sprint"
                ],
                "summary": "Sprint creation",
                "description": "Returns created Sprint",
                "operationId": "create-sprint",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "authorization header",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "sprintTitle",
                        "in": "formData",
                        "description": "Sprint Title",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "startDate",
                        "in": "formData",
                        "description": "start datetime",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "endDate",
                        "in": "formData",
                        "description": "End Datetime",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "status",
                        "in": "formData",
                        "description": "Sprint Status('created', 'assigned', 'onhold', 'inprogress','completed', 'cancelled',' failed')",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "priority",
                        "in": "formData",
                        "description": "Sprint Priority('critical', 'high', 'medium', 'low')",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "task_id",
                        "in": "formData",
                        "description": "Id of the Task",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "name": "dependent_sprint_id",
                        "in": "formData",
                        "description": "Dependent Sprint ID",
                        "required": true,
                        "type": "number"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "500": {
                        "description": "Internal server error"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/v1/sprint/{id}": {
            "get": {
                "tags": [
                    "Sprint"
                ],
                "summary": "Sprint details",
                "description": "Returns Sprint details",
                "operationId": "single-Sprint",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "authorization header",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Sprint Id",
                        "required": true,
                        "type": "number"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "500": {
                        "description": "Internal server error"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            },
            "put": {
                "tags": [
                    "Sprint"
                ],
                "summary": "Sprint updation",
                "description": "Returns updated Sprint",
                "operationId": "update-sprint",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "authorization header",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "id",
                        "in": "formData",
                        "description": "id of the Sprint",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "name": "sprintTitle",
                        "in": "formData",
                        "description": "Sprint Title",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "startDate",
                        "in": "formData",
                        "description": "start datetime",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "endDate",
                        "in": "formData",
                        "description": "End Datetime",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "status",
                        "in": "formData",
                        "description": "Sprint Status('created', 'assigned', 'onhold', 'inprogress','completed', 'cancelled',' failed')",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "priority",
                        "in": "formData",
                        "description": "Sprint Priority('critical', 'high', 'medium', 'low')",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "task_id",
                        "in": "formData",
                        "description": "Id of the Task",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "name": "sprint_assigned_by",
                        "in": "formData",
                        "description": "Id of the User",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "name": "sprint_handled_by",
                        "in": "formData",
                        "description": "Id of the User",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "name": "dependent_sprint_id",
                        "in": "formData",
                        "description": "Dependent Sprint ID",
                        "required": true,
                        "type": "number"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "500": {
                        "description": "Internal server error"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Sprint"
                ],
                "summary": "Delete a Sprint",
                "description": "Delete a Sprint",
                "operationId": "delete Sprint",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "authorization header",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Sprint Id",
                        "required": true,
                        "type": "number"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "500": {
                        "description": "Internal server error"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/v1/sprint/by-milestone/{id}": {
            "get": {
                "tags": [
                    "Sprint"
                ],
                "summary": "Dependent Sprint details from milestone",
                "description": "Returns Dependent Sprint details",
                "operationId": "Dependent-Sprint-Details",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "authorization header",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "milestone Id",
                        "required": true,
                        "type": "number"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "500": {
                        "description": "Internal server error"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
...
    },
...
}